refactor(agents): drop underscore prefix from skills parsing module
