+++ src/phoenix/server/agents/capabilities/skills/__init__.py
@@ -1,5 +1,5 @@
-from phoenix.server.agents.capabilities.skills._parsing import parse_skill_md
 from phoenix.server.agents.capabilities.skills.capability import SkillsCapability
+from phoenix.server.agents.capabilities.skills.parsing import parse_skill_md
 from phoenix.server.agents.capabilities.skills.toolset import SkillsToolset
 from phoenix.server.agents.capabilities.skills.types import Skill, SkillResource
